qcacld-3.0: Add host support for db2dbm RSSI changes
Add host support for db2dbm RSSI changes. Firmware
indicates this capability when underlying hardware
has RSSI reporting feature. Based on this capability
host will know if firmware sends SNR or RSSI. If no
capablity is present then host will convert SNR to
rssi using a fixed offset of -96. If capability is
present host will directly use the rssi as it is.

CRs-Fixed: 2364023

+++ core/sme/inc/csr_api.h
@@ -1293,7 +1293,7 @@ typedef struct sSirSmeAssocIndToUpperLayerCnf {
 
 typedef struct tagCsrSummaryStatsInfo {
 	uint32_t snr;
-	uint32_t rssi;
+	int8_t rssi;
 	uint32_t retry_cnt[4];
 	uint32_t multiple_retry_cnt[4];
 	uint32_t tx_frm_cnt[4];
